+++ Rotation/PersephoneCore/Context/IPersephoneContext.cs
@@ -1,5 +1,6 @@
 using Olympus.Rotation.Common;
 using Olympus.Rotation.PersephoneCore.Helpers;
+using Olympus.Services.Party;
 
 namespace Olympus.Rotation.PersephoneCore.Context;
 
@@ -239,6 +240,16 @@ public interface IPersephoneContext : ICasterDpsRotationContext
     /// Debug state for this rotation.
     /// </summary>
     PersephoneDebugState Debug { get; }
+
+    #endregion
+
+    #region Party Coordination
+
+    /// <summary>
+    /// Service for coordinating raid buffs with other Olympus instances.
+    /// Null if party coordination is disabled or unavailable.
+    /// </summary>
+    IPartyCoordinationService? PartyCoordinationService { get; }
 
     #endregion
 }